Bug fix pir
