+++ firmware/ezlopi/devices/digital_io/CMakeLists.txt
@@ -3,6 +3,6 @@
 file(GLOB_RECURSE config_src "*.c")
 # set(config_src)
 
-idf_component_register(SRCS "digital_io_isr.c" "digital_io.c" "${config_src}"
+idf_component_register(SRCS "digital_io.c" "${config_src}"
     INCLUDE_DIRS "."
-    REQUIRES trace library ezlopi_actions  ezlopi_timer ezlopi_devices ezlopi_cloud_constants)
+    REQUIRES trace library ezlopi_actions  ezlopi_timer ezlopi_devices ezlopi_cloud_constants gpio_isr_service)
