+++ firmware/ezlopi/devices/sensor_pir/CMakeLists.txt
@@ -3,4 +3,4 @@
 
 idf_component_register(SRCS "sensor_pir.c"
                         INCLUDE_DIRS .
-                        REQUIRES trace ezlopi_actions ezlopi_timer items isr_service)
+                        REQUIRES trace ezlopi_actions ezlopi_timer items gpio_isr_service)
