+++ firmware/ezlopi/main/main.c
@@ -18,9 +18,7 @@
 #include "qt_serial.h"
 #include "web_provisioning.h"
 #include "gatt_server.h"
-#include "sensor_pir.h"
-
-#define EZLOPI_PIR_PIN_NUM GPIO_NUM_2
+#include "gpio_isr_service.h"
 
 static void blinky(void *pv);
 
@@ -34,6 +32,7 @@ void app_main(void)
     web_provisioning_init();
     GATT_SERVER_MAIN();
     sensor_service_init();
+    gpio_isr_service_init();
 
     // xTaskCreate(blinky, "blinky", 2 * 2048, NULL, 1, NULL);
 }
