+++ firmware/ezlopi/services/gpio_isr_service/component.mk
@@ -1,0 +1,8 @@
+#
+# "main" pseudo-component makefile.
+#
+# (Uses default behaviour of compiling all source files in directory, adding 'include' to include path.)
+
+
+COMPONENT_ADD_SRCDIRS := .
+COMPONENT_ADD_INCLUDEDIRS := .
